requirements.txt
absl-py==1.4.0
ansicon==1.89.0
appdirs==1.4.4
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
AudioConverter==1.0.0
audioread==3.0.0
backcall==0.2.0
blessed==1.20.0
blinker==1.6.2
Brotli==1.0.9
build==0.10.0
cachecontrol==0.13.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cleo==2.0.1
click==8.1.7
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
crashtest==0.4.1
cssselect2==0.7.0
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.7
dulwich==0.21.6
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.3
Flask==2.2.3
Flask-Mail==0.9.1
flatbuffers==23.1.21
flyingcircus==0.1.4.1
fonttools==4.38.0
fst-pso==1.8.1
funcy==1.18
FuzzyTM==2.0.5
gast==0.4.0
gensim==4.3.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
helper==2.5.0
html5lib==1.1
idna==3.4
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
installer==0.7.0
ipykernel==6.21.2
ipython==8.10.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.18.2
Jinja2==3.1.2
jinxed==1.2.0
joblib==1.2.0
jsonschema==4.17.3
jupyter-client==8.0.3
jupyter-core==5.2.0
keras==2.10.0
Keras-Preprocessing==1.1.2
keyring==24.2.0
kiwisolver==1.4.4
lazy-loader==0.1
libclang==15.0.6.1
librosa==0.10.0
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
miniful==0.0.6
more-itertools==10.1.0
msgpack==1.0.4
nest-asyncio==1.5.6
nltk==3.8.1
numba==0.56.4
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.1.1
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pdfkit==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkginfo==1.9.6
platformdirs==3.0.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
pooch==1.6.0
prompt-toolkit==3.0.37
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycparser==2.21
pydub==0.25.1
pydyf==0.7.0
pyFUME==0.2.25
Pygments==2.14.0
pyLDAvis==3.4.0
pyparsing==3.0.9
pyphen==0.14.0
pyproject-hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytk==0.0.2.1
pytz==2022.7.1
pywin32==305
pywin32-ctypes==0.2.2
PyYAML==6.0
pyzmq==25.0.0
rapidfuzz==2.15.1
regex==2022.10.31
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
setuptools-scm==7.1.0
shellingham==1.5.3
simpful==2.10.0
six==1.16.0
smart-open==6.3.0
smtp-mail==0.0.2
soundfile==0.12.1
soxr==0.3.3
stack-data==0.6.2
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.64.1
traitlets==5.9.0
trove-classifiers==2023.8.7
typing-extensions==4.7.1
urllib3==1.26.14
virtualenv==20.24.4
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
Werkzeug==2.2.3
wkhtmltopdf==0.2
wordcloud==1.9.1.1
wrapt==1.14.1
zipp==3.15.0
zopfli==0.2.2
